Determine `active_server` once per cluster, not once for all
While `first_server.yml` would only be included for the 1st server of a cluster,
`run_once` for setting the `active_server` fact also meant this would only happen once
for all clusters instead of once per cluster.

Use a loop across the members of each cluster and a fact delegation instead, to set the
`active_server` fact instead on each cluster member to the `inventory_hostname` of the
server, for which `first_server.yml` was included.