Add additional upgrade options
Add `rke2_wait_for_all_pods_to_be_healthy` and
`rke2_wait_for_all_pods_to_be_healthy_args`
Currently, the `Wait for all pods to be ready again` task does only
checks that pods are in the running or succeeded phase, it does not
check they are "ready" and passing their health checks. THere is now a
new task that uses kubectl wait to verify the pods are ready/passing
their health checks. To maintain compatibility with current behaviour
this is flagged default off with a new var.

Add `rke2_drain_additional_args` to allow args such as --pod-selector to
be passed to the drain command. This is helpful for when certain pods
such as longhorn-instance-manager don't drain properly.